# Conflicts:
#	ibs-ui/src/main/web/src/app/admin/admin.module.ts
#	ibs-ui/src/main/web/src/app/admin/book-edit/book-edit.component.ts

+++ ibs-db-services/src/main/java/ru/ibs/db/service/beans/AuthorService.java
@@ -5,7 +5,7 @@ import ru.ibs.entity.Author;
 import ru.ibs.repository.AuthorRepository;
 
 @Service
-public class AuthorService extends JpaService<Author, Integer> {
+public class AuthorService extends JpaService<Author, Long> {
 
     private final AuthorRepository repository;
